Renaming gcm feature.py module to feature_relevance.py
The new name 'feature_relevance' will help clarify the type of functionality to expect from the module.
